requirements.txt
fastapi
uvicorn
pymongo
pandas
requests
httpx
pytest
pydantic-settings